Start of modeling the enclosure with OpenSCAD.
So far we have a model of the enclosure itself (which is a product from
Polycase) and a model of what turned out to be the wrong size of button.
There's also an overly-detailed STL model of the Raspberry Pi A+ in there
which we'll probably later replace with a simpler model that's more
amenable to CSG operations.