Add dummy copper layer to fake out signal connections on header pin.
In order for the DRC to consider the routing to be complete it needs to
know that the GND, VCC and AVCC pins (respectively) on the Raspberry Pi
header will be connected by the Pi board itself and thus don't need to
actually be connected on this board. We achieve this by providing a dummy
copper layer that won't be included in the fab order.